refactor: move organization to service layer
